Correction on betty style

+++ 0x0F-function_pointers/1-array_iterator.c
@@ -2,7 +2,7 @@
 #include "function_pointers.h"
 
 /**
- * array_iterator - executes a function given as a parameter 
+ * array_iterator - executes a function given as a parameter
  * on each element of an array
  * @array: array to iterate over
  * @size: size of the array
@@ -10,7 +10,7 @@
  */
 void array_iterator(int *array, size_t size, void (*action)(int))
 {
-	size_t i;
+	unsigned int i;
 
 	if (!array || !action)
 		return;
